% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_colours.R
\name{diagram_colours}
\alias{diagram_colours}
\title{Available colours in package}
\usage{
diagram_colours(palette = "", n_colors = 10, n_colours = n_colors)
}
\arguments{
\item{palette}{(optional) a string or vector of strings.
Either the palette name if using a predefined colour scheme,
or a vector or hexidecimal colours if using own colour scheme.
Returns a vector of hexidecimal values, or NULL if empty (default colour scheme for plotly)}

\item{n_colors}{(optional) localisation of n_colours}

\item{n_colours}{(optional) the number of hexidecimal values to generate
standard is 10 as that is the maximum expected tissue types}
}
\value{
Returns a vector of hexidecimal colours, or NULL if left blank
}
\description{
Outputs a vector of hexidecimal colour values if the colour palette is in
hcl.pals or viridis.

Used to validate the colours that will output in these packages to help
Create custom palettes. Advised to output 10 colours, as this is the maximum
number of tissue types.
}
\details{
Valid palette inputs:
\itemize{
\item Viridis options: \code{"viridis"}, \code{"magma"}, \code{"plasma"},
\code{"inferno"}, \code{"cividis"}, \code{"mako"}, \code{"rocket"}, \code{"turbo"}.
\item Any palette name returned by \code{grDevices::hcl.pals()}.
}

To see available HCL palettes, run \code{grDevices::hcl.pals()}.
}
\examples{
diagram_colours("Reds")

}
