% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinviz.R
\docType{package}
\name{spinviz}
\alias{spinviz-package}
\alias{spinviz}
\title{spinviz: Injury visualisation tools}
\description{
Tools for rendering injury frequency heatmaps on human-body SVG diagrams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bnqcasimiro/spinviz}
  \item Report bugs at \url{https://github.com/bnqcasimiro/spinviz/issues}
}

}
\author{
\strong{Maintainer}: Bianca Nicole Casimiro \email{biancanicole.casimiro@students.mq.edu.au}

Authors:
\itemize{
  \item Bianca Nicole Casimiro \email{biancanicole.casimiro@students.mq.edu.au}
  \item Thomas Fung \email{thomas.fung.dr@gmail.com}
  \item Reidar Lystad \email{reidar.lystad@mq.edu.au}
}

}
