% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_diagram.R
\name{injury_heatmap}
\alias{injury_heatmap}
\title{Heat map of Injuries on Human Body}
\usage{
injury_heatmap(
  injury_data,
  selected_sport,
  view_choice,
  sex = c("male", "female"),
  palette = NULL,
  opacity = 0.8,
  show_labels = TRUE,
  show_values = TRUE,
  show_scale = TRUE
)
}
\arguments{
\item{injury_data}{a data frame with the body region, area the injury occurred,
and the frequency of the injury}

\item{selected_sport}{the name of the sport, corresponding to the header of
the frequency column in the data frame.}

\item{view_choice}{which view to render the diagram. One of ("front", "back", or "both")}

\item{sex}{which body diagram to use: \code{"male"} or \code{"female"}.}

\item{palette}{WIP - placeholder to add custom colour options}

\item{opacity}{Numeric between \code{0} and \code{1}. Opacity of the filled body regions.}

\item{show_labels}{Logical. If \code{TRUE}, show region names.}

\item{show_values}{(optional) Logical. If \code{TRUE}, show injury values.}

\item{show_scale}{(optional) Logical. If \code{TRUE}, show the colour scale (legend).}
}
\value{
A plot in R-studio viewer
}
\description{
Creates a heat map projected onto a diagram of the human body,
based on the number of injuries on body parts visible on the
diagram.
Diagrams can be viewed from the front, back, or both views
}
\examples{
Subcategory <- c("Head","Neck","Shoulder","Chest","Upper Arm","Elbow",
                  "Abdomen","Forearm","Hip Groin","Wrist","Hand",
                  "Thigh","Knee","Lower Leg","Ankle","Foot","Thoracic Spine","Lumbosacral")
Region.area <- rep("Example", length(Subcategory))
boxing <- c(15, 5, 18, 12, 20, 6, 10, 14, 9, 9, 11, 3, 16, 13, 7, 8, 18, 22)
df <- data.frame(Region.area, Subcategory, boxing)
# Generate a plot for front view, male
p <- injury_heatmap(df, "boxing", "front", sex = "male", show_values = FALSE)
# You can add your own plot title by:
p + ggplot2::labs(
  title = "Boxing Injury Heatmap (Front)"
)
}
