% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinviz.R
\docType{package}
\name{spinviz}
\alias{spinviz-package}
\alias{spinviz}
\title{spinviz: Injury visualisation tools}
\description{
Tools for rendering injury frequency heatmaps on human-body SVG diagrams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bnqcasimiro/spinviz}
  \item Report bugs at \url{https://github.com/bnqcasimiro/spinviz/issues}
}

}
\author{
\strong{Maintainer}: Bianca Nicole Casimiro \email{biancanicole.casimiro@students.mq.edu.au}

Authors:
\itemize{
  \item Bianca Nicole Casimiro \email{biancanicole.casimiro@students.mq.edu.au}
  \item Thomas Fung \email{thomas.fung.dr@gmail.com}
  \item Reidar Lystad \email{reidar.lystad@mq.edu.au}
}

Other contributors:
\itemize{
  \item Alexander Brinkman \email{alexander.brinkman@students.mq.edu.au} [contributor]
  \item Alexandra Dooley \email{alexandra.dooley@students.mq.edu.au} [contributor]
  \item Andisheh Saffarian \email{andisheh.saffarian@students.mq.edu.au} [contributor]
  \item Brice Thu \email{brice.thu1@students.mq.edu.au} [contributor]
  \item Utsav Chadha \email{utsav.chadha@students.mq.edu.au} [contributor]
  \item Zoe Remo \email{zoe.remo@students.mq.edu.au} [contributor]
  \item Brayden Smith \email{brayden.smith@students.mq.edu.au} [contributor]
  \item Govardhan Bharadwaj \email{govardhan.bharadwaj@students.mq.edu.au} [contributor]
  \item Katja Amet \email{katja.amet@students.mq.edu.au} [contributor]
  \item Kyle Mcnicholas \email{kyle.mcnicholas@students.mq.edu.au} [contributor]
  \item Zachary Carr \email{zachary.carr@students.mq.edu.au} [contributor]
}

}
